+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/FederationGatewaySyncService.java
@@ -3,6 +3,10 @@ package fi.thl.covid19.exposurenotification.efgs;
 import fi.thl.covid19.exposurenotification.diagnosiskey.DiagnosisKeyDao;
 import fi.thl.covid19.exposurenotification.diagnosiskey.IntervalNumber;
 import fi.thl.covid19.exposurenotification.diagnosiskey.TemporaryExposureKey;
+import fi.thl.covid19.exposurenotification.efgs.entity.DownloadData;
+import fi.thl.covid19.exposurenotification.efgs.entity.FederationOutboundOperation;
+import fi.thl.covid19.exposurenotification.efgs.entity.UploadResponseEntity;
+import fi.thl.covid19.exposurenotification.efgs.signing.FederationGatewaySigning;
 import fi.thl.covid19.proto.EfgsProto;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
@@ -16,8 +20,8 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static fi.thl.covid19.exposurenotification.diagnosiskey.DiagnosisKeyDao.MAX_RETRY_COUNT;
-import static fi.thl.covid19.exposurenotification.efgs.FederationGatewayBatchSignatureUtil.validateSignature;
-import static fi.thl.covid19.exposurenotification.efgs.FederationGatewayBatchUtil.*;
+import static fi.thl.covid19.exposurenotification.efgs.util.SignatureValidationUtil.validateSignature;
+import static fi.thl.covid19.exposurenotification.efgs.util.BatchUtil.*;
 import static fi.thl.covid19.exposurenotification.efgs.OperationDao.EfgsOperationDirection.*;
 import static java.util.Objects.requireNonNull;
 
@@ -26,13 +30,13 @@ public class FederationGatewaySyncService {
     private final FederationGatewayClient client;
     private final DiagnosisKeyDao diagnosisKeyDao;
     private final OperationDao operationDao;
-    private final FederationGatewayBatchSigner signer;
+    private final FederationGatewaySigning signer;
 
     public FederationGatewaySyncService(
             FederationGatewayClient client,
             DiagnosisKeyDao diagnosisKeyDao,
             OperationDao operationDao,
-            FederationGatewayBatchSigner signer
+            FederationGatewaySigning signer
     ) {
         this.client = requireNonNull(client);
         this.diagnosisKeyDao = requireNonNull(diagnosisKeyDao);
