+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/entity/UploadResponseEntity.java
@@ -1,4 +1,4 @@
-package fi.thl.covid19.exposurenotification.efgs;
+package fi.thl.covid19.exposurenotification.efgs.entity;
 
 import org.springframework.http.HttpStatus;
