+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/scheduled/CallbackInitializer.java
@@ -1,5 +1,7 @@
-package fi.thl.covid19.exposurenotification.efgs;
+package fi.thl.covid19.exposurenotification.efgs.scheduled;
 
+import fi.thl.covid19.exposurenotification.efgs.FederationGatewayClient;
+import fi.thl.covid19.exposurenotification.efgs.entity.Callback;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
