+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/util/DsosMapperUtil.java
@@ -1,4 +1,4 @@
-package fi.thl.covid19.exposurenotification.efgs;
+package fi.thl.covid19.exposurenotification.efgs.util;
 
 import java.util.Arrays;
 import java.util.Optional;
