+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/efgs/util/SignatureValidationUtil.java
@@ -1,7 +1,7 @@
-package fi.thl.covid19.exposurenotification.efgs;
+package fi.thl.covid19.exposurenotification.efgs.util;
 
-import com.google.protobuf.ByteString;
-import com.google.protobuf.ProtocolStringList;
+import fi.thl.covid19.exposurenotification.efgs.entity.AuditEntry;
+import fi.thl.covid19.exposurenotification.efgs.entity.DownloadData;
 import fi.thl.covid19.proto.EfgsProto;
 import org.bouncycastle.asn1.x500.style.BCStyle;
 import org.bouncycastle.cert.X509CertificateHolder;
@@ -12,27 +12,25 @@ import org.bouncycastle.operator.OperatorCreationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.StringReader;
-import java.math.BigInteger;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
 import java.security.*;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import static fi.thl.covid19.exposurenotification.efgs.util.SignatureHelperUtil.generateBytesForSignature;
+import static fi.thl.covid19.exposurenotification.efgs.util.SignatureHelperUtil.getCertThumbprint;
 import static net.logstash.logback.argument.StructuredArguments.keyValue;
 
 /*
  * Some parts are strictly based on efgs implementation to achieve compability.
  * See: https://github.com/eu-federation-gateway-service/efgs-federation-gateway/tree/master/src/main/java/eu/interop/federationgateway/batchsigning
  */
-public class FederationGatewayBatchSignatureUtil {
+public class SignatureValidationUtil {
 
-    private static final Logger LOG = LoggerFactory.getLogger(FederationGatewayBatchSignatureUtil.class);
+    private static final Logger LOG = LoggerFactory.getLogger(SignatureValidationUtil.class);
 
     public static EfgsProto.DiagnosisKeyBatch validateSignature(
             List<AuditEntry> auditEntries, DownloadData downloadData, X509Certificate trustAnchor) {
@@ -83,7 +81,7 @@ public class FederationGatewayBatchSignatureUtil {
     }
 
     private static boolean verifySignedDataCertificate(CMSSignedData signedData, SignerInformation signerInfo, AuditEntry audit)
-    throws IOException, NoSuchAlgorithmException {
+            throws IOException, NoSuchAlgorithmException {
         X509CertificateHolder certFromSignedData = (X509CertificateHolder) signedData.getCertificates().getMatches(signerInfo.getSID()).iterator().next();
         return certFromSignedData.isValidOn(new Date()) &&
                 getCertThumbprint(certFromSignedData).equals(audit.uploaderSigningThumbprint);
@@ -97,97 +95,6 @@ public class FederationGatewayBatchSignatureUtil {
     private static SignerInformationVerifier createSignerInfoVerifier(X509CertificateHolder signerCert)
             throws OperatorCreationException, CertificateException {
         return new JcaSimpleSignerInfoVerifierBuilder().build(signerCert);
-    }
-
-    public static byte[] generateBytesForSignature(List<EfgsProto.DiagnosisKey> keys) {
-        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-
-        keys.stream()
-                .map(FederationGatewayBatchSignatureUtil::generateBytesToVerify)
-                .sorted(Comparator.nullsLast(
-                        Comparator.comparing(FederationGatewayBatchSignatureUtil::bytesToBase64)
-                ))
-                .forEach(byteArrayOutputStream::writeBytes);
-
-        return byteArrayOutputStream.toByteArray();
-    }
-
-    private static byte[] generateBytesToVerify(EfgsProto.DiagnosisKey diagnosisKey) {
-        final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-        writeBytesInByteArray(diagnosisKey.getKeyData(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeIntInByteArray(diagnosisKey.getRollingStartIntervalNumber(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeIntInByteArray(diagnosisKey.getRollingPeriod(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeIntInByteArray(diagnosisKey.getTransmissionRiskLevel(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeVisitedCountriesInByteArray(diagnosisKey.getVisitedCountriesList(),
-                byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeB64StringInByteArray(diagnosisKey.getOrigin(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeIntInByteArray(diagnosisKey.getReportTypeValue(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        writeIntInByteArray(diagnosisKey.getDaysSinceOnsetOfSymptoms(), byteArrayOutputStream);
-        writeSeperatorInArray(byteArrayOutputStream);
-        return byteArrayOutputStream.toByteArray();
-    }
-
-    public static String bytesToBase64(byte[] bytes) {
-        return Base64.getEncoder().encodeToString(bytes);
-    }
-
-    private static void writeSeperatorInArray(final ByteArrayOutputStream byteArray) {
-        byteArray.writeBytes(".".getBytes(StandardCharsets.US_ASCII));
-    }
-
-    private static void writeStringInByteArray(final String batchString, final ByteArrayOutputStream byteArray) {
-        byteArray.writeBytes(batchString.getBytes(StandardCharsets.US_ASCII));
-    }
-
-    private static void writeB64StringInByteArray(final String batchString, final ByteArrayOutputStream byteArray) {
-        String base64String = bytesToBase64(batchString.getBytes(StandardCharsets.US_ASCII));
-
-        if (base64String != null) {
-            writeStringInByteArray(base64String, byteArray);
-        }
-    }
-
-    private static void writeIntInByteArray(final int batchInt, final ByteArrayOutputStream byteArray) {
-        String base64String = bytesToBase64(ByteBuffer.allocate(4).putInt(batchInt).array());
-
-        if (base64String != null) {
-            writeStringInByteArray(base64String, byteArray);
-        }
-    }
-
-    private static void writeBytesInByteArray(final ByteString bytes, ByteArrayOutputStream byteArray) {
-        String base64String = bytesToBase64(bytes.toByteArray());
-
-        if (base64String != null) {
-            writeStringInByteArray(base64String, byteArray);
-        }
-    }
-
-    private static void writeVisitedCountriesInByteArray(final ProtocolStringList countries,
-                                                         final ByteArrayOutputStream byteArray) {
-        writeB64StringInByteArray(String.join(",", countries), byteArray);
-    }
-
-    public static String getCertThumbprint(X509CertificateHolder x509CertificateHolder) throws IOException, NoSuchAlgorithmException {
-        return calculateHash(x509CertificateHolder.getEncoded());
-    }
-
-    private static String calculateHash(byte[] data) throws NoSuchAlgorithmException {
-        byte[] certHashBytes = MessageDigest.getInstance("SHA-256").digest(data);
-        String hexString = new BigInteger(1, certHashBytes).toString(16);
-
-        if (hexString.length() == 63) {
-            hexString = "0" + hexString;
-        }
-
-        return hexString;
     }
 
     private static boolean verifyOperatorSignature(AuditEntry audit, X509Certificate trustAnchor)
