+++ exposure-notification/src/test/java/fi/thl/covid19/exposurenotification/efgs/CallbackInitializerTestIT.java
@@ -2,6 +2,8 @@ package fi.thl.covid19.exposurenotification.efgs;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import fi.thl.covid19.exposurenotification.efgs.entity.Callback;
+import fi.thl.covid19.exposurenotification.efgs.scheduled.CallbackInitializer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
